% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_files.R
\name{api_server_available}
\alias{api_server_available}
\title{Check if SWAT+ Editor API server is available}
\usage{
api_server_available(api_url)
}
\arguments{
\item{api_url}{API base URL.}
}
\value{
Logical. TRUE if server responds.
}
\description{
Check if SWAT+ Editor API server is available
}
\keyword{internal}
