% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_swatplus.R
\name{check_swatplus}
\alias{check_swatplus}
\title{Check SWAT+ input file consistency}
\usage{
check_swatplus(project, output_dir)
}
\arguments{
\item{project}{List. A SWAT+ project object with \code{db_file}.}

\item{output_dir}{Character. Directory where the input files will be
written. Defaults to \code{"TxtInOut"} in the project directory.}
}
\value{
A list of issues found, or an empty list if no issues detected. Each issue is a character string describing the problem.
}
\description{
Check SWAT+ input file consistency
}
