% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connect.R
\name{close_db}
\alias{close_db}
\title{Safely close a database connection}
\usage{
close_db(con)
}
\arguments{
\item{con}{A DBI connection object.}
}
\description{
Safely close a database connection
}
\keyword{internal}
