% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_files.R
\name{copy_weather_files}
\alias{copy_weather_files}
\title{Copy weather data files from weather_dir to output_dir}
\usage{
copy_weather_files(
  con,
  output_dir,
  weather_dir,
  weather_data_format = "observed"
)
}
\description{
Copy weather data files from weather_dir to output_dir
}
\keyword{internal}
