% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connect.R
\name{execute_db}
\alias{execute_db}
\title{Execute a statement on the project database (INSERT, UPDATE, DELETE)}
\usage{
execute_db(con, statement, params = NULL)
}
\arguments{
\item{con}{A DBI connection object.}

\item{statement}{Character. SQL statement to execute.}

\item{params}{List. Optional statement parameters.}
}
\value{
Number of affected rows.
}
\description{
Execute a statement on the project database (INSERT, UPDATE, DELETE)
}
\keyword{internal}
