% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_closest_station}
\alias{find_closest_station}
\title{Find the closest station by latitude/longitude}
\usage{
find_closest_station(stations, lat, lon)
}
\arguments{
\item{stations}{A data.frame with lat and lon columns.}

\item{lat}{Numeric. Target latitude.}

\item{lon}{Numeric. Target longitude.}
}
\value{
Integer. Row index of the closest station.
}
\description{
Find the closest station by latitude/longitude
}
\keyword{internal}
