% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileio.R
\name{get_cio_file_names}
\alias{get_cio_file_names}
\title{Get file names from file_cio for a classification section}
\usage{
get_cio_file_names(con, section)
}
\arguments{
\item{con}{DBI connection.}

\item{section}{Character. Classification name (e.g. "simulation").}
}
\value{
Character vector of file names.
}
\description{
Get file names from file_cio for a classification section
}
\keyword{internal}
