% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_files.R
\name{get_file_cio_conditions}
\alias{get_file_cio_conditions}
\title{Get file.cio condition flags for each classification}
\usage{
get_file_cio_conditions(con, is_lte = FALSE, is_netcdf = FALSE)
}
\description{
Get file.cio condition flags for each classification
}
\keyword{internal}
