% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwflow.R
\name{get_gwflow_base}
\alias{get_gwflow_base}
\title{Get GWFLOW base configuration}
\usage{
get_gwflow_base(project)
}
\arguments{
\item{project}{List. A SWAT+ project object with \code{db_file}.}
}
\value{
A list with GWFLOW base settings.
}
\description{
Get GWFLOW base configuration
}
