% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwflow.R
\name{get_gwflow_zones}
\alias{get_gwflow_zones}
\title{Get GWFLOW zones}
\usage{
get_gwflow_zones(project)
}
\arguments{
\item{project}{List. A SWAT+ project object with \code{db_file}.}
}
\value{
A data.frame with zone parameters.
}
\description{
Get GWFLOW zones
}
