% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{get_project_config}
\alias{get_project_config}
\title{Get project configuration from the database}
\usage{
get_project_config(project)
}
\arguments{
\item{project}{List. A SWAT+ project object with \code{db_file}.}
}
\value{
A list with project configuration values.
}
\description{
Get project configuration from the database
}
\examples{
\dontrun{
config <- get_project_config(project)
}
}
