% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_files.R
\name{gwflow_exists}
\alias{gwflow_exists}
\title{Check if gwflow module is active}
\usage{
gwflow_exists(con)
}
\arguments{
\item{con}{Database connection.}
}
\value{
Logical. TRUE if gwflow_base table exists and has data,
AND project_config.use_gwflow is enabled.
}
\description{
Check if gwflow module is active
}
\keyword{internal}
