% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_files.R
\name{gwflow_gis_to_con_index}
\alias{gwflow_gis_to_con_index}
\title{Build GIS ID to connection index mapping}
\usage{
gwflow_gis_to_con_index(con, con_table)
}
\description{
Build GIS ID to connection index mapping
}
\keyword{internal}
