% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_files.R
\name{gwflow_grid_index}
\alias{gwflow_grid_index}
\title{Build grid index mapping cell_id -> cell data}
\usage{
gwflow_grid_index(con)
}
\description{
Build grid index mapping cell_id -> cell data
}
\keyword{internal}
