% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_files.R
\name{gwflow_write_grid_column}
\alias{gwflow_write_grid_column}
\title{Write a grid column for gwflow_init_conc}
\usage{
gwflow_write_grid_column(con, gw, column_name)
}
\description{
Write a grid column for gwflow_init_conc
}
\keyword{internal}
