% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileio.R
\name{has_data}
\alias{has_data}
\title{Safely check if a table exists and has data}
\usage{
has_data(con, table_name)
}
\arguments{
\item{con}{DBI connection.}

\item{table_name}{Table name.}
}
\value{
Logical.
}
\description{
Safely check if a table exists and has data
}
\keyword{internal}
