% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_nonempty_string}
\alias{is_nonempty_string}
\title{Check if a value is a non-empty string}
\usage{
is_nonempty_string(x)
}
\arguments{
\item{x}{Value to check.}
}
\value{
Logical.
}
\description{
Check if a value is a non-empty string
}
\keyword{internal}
