% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connect.R
\name{list_db_tables}
\alias{list_db_tables}
\title{List all tables in the project database}
\usage{
list_db_tables(con)
}
\arguments{
\item{con}{A DBI connection object.}
}
\value{
Character vector of table names.
}
\description{
List all tables in the project database
}
\keyword{internal}
