% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{list_weather_stations}
\alias{list_weather_stations}
\title{List weather stations in the project database}
\usage{
list_weather_stations(project)
}
\arguments{
\item{project}{List. A SWAT+ project object with \code{db_file}.}
}
\value{
A data.frame with all weather stations.
}
\description{
List weather stations in the project database
}
\examples{
\dontrun{
stations <- list_weather_stations(project)
}
}
