% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connect.R
\name{open_project_db}
\alias{open_project_db}
\title{Open a connection to the SWAT+ project database}
\usage{
open_project_db(db_path)
}
\arguments{
\item{db_path}{Character. Path to the SQLite project database file.}
}
\value{
A DBI connection object.
}
\description{
Open a connection to the SWAT+ project database
}
\examples{
\dontrun{
con <- open_project_db("/path/to/project.sqlite")
DBI::dbDisconnect(con)
}
}
