% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{populate_gis_from_project}
\alias{populate_gis_from_project}
\title{Populate GIS tables from project basin/HRU data}
\usage{
populate_gis_from_project(con, project)
}
\arguments{
\item{con}{DBI connection.}

\item{project}{Project list object.}
}
\description{
Populate GIS tables from project basin/HRU data
}
\keyword{internal}
