% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connect.R
\name{read_db_table}
\alias{read_db_table}
\title{Read a table from the project database}
\usage{
read_db_table(con, table_name)
}
\arguments{
\item{con}{A DBI connection object.}

\item{table_name}{Character. Name of the table to read.}
}
\value{
A data.frame with the table contents.
}
\description{
Read a table from the project database
}
\keyword{internal}
