% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_data.R
\name{read_gis_aquifers}
\alias{read_gis_aquifers}
\title{Read GIS aquifers table}
\usage{
read_gis_aquifers(project)
}
\arguments{
\item{project}{List. A SWAT+ project object with \code{db_file}.}
}
\value{
A data.frame with aquifer data.
}
\description{
Read GIS aquifers table
}
