% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_data.R
\name{read_gis_channels}
\alias{read_gis_channels}
\title{Read GIS channels table}
\usage{
read_gis_channels(project)
}
\arguments{
\item{project}{List. A SWAT+ project object with \code{db_file}.}
}
\value{
A data.frame with channel data (length, slope, width, depth).
}
\description{
Read GIS channels table
}
