% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_data.R
\name{read_gis_data}
\alias{read_gis_data}
\title{Read all GIS data from the project database}
\usage{
read_gis_data(project)
}
\arguments{
\item{project}{List. A SWAT+ project object with \code{db_file}.}
}
\value{
A named list of data.frames for each GIS table present in the
database.
}
\description{
Returns a list containing all available GIS tables as data.frames.
}
\examples{
\dontrun{
gis <- read_gis_data(project)
names(gis)
head(gis$subbasins)
}
}
