% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_data.R
\name{read_gis_hrus}
\alias{read_gis_hrus}
\title{Read GIS HRUs table}
\usage{
read_gis_hrus(project)
}
\arguments{
\item{project}{List. A SWAT+ project object with \code{db_file}.}
}
\value{
A data.frame with HRU data (landuse, soil, slope, area).
}
\description{
Read GIS HRUs table
}
