% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_data.R
\name{read_gis_lsus}
\alias{read_gis_lsus}
\title{Read GIS landscape units table}
\usage{
read_gis_lsus(project)
}
\arguments{
\item{project}{List. A SWAT+ project object with \code{db_file}.}
}
\value{
A data.frame with LSU data.
}
\description{
Read GIS landscape units table
}
