% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_data.R
\name{read_gis_points}
\alias{read_gis_points}
\title{Read GIS points table}
\usage{
read_gis_points(project)
}
\arguments{
\item{project}{List. A SWAT+ project object with \code{db_file}.}
}
\value{
A data.frame with point feature data.
}
\description{
Read GIS points table
}
