% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_data.R
\name{read_gis_routing}
\alias{read_gis_routing}
\title{Read GIS routing table}
\usage{
read_gis_routing(project)
}
\arguments{
\item{project}{List. A SWAT+ project object with \code{db_file}.}
}
\value{
A data.frame with routing connectivity data.
}
\description{
Read GIS routing table
}
