% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_data.R
\name{read_gis_subbasins}
\alias{read_gis_subbasins}
\title{Read GIS subbasins table}
\usage{
read_gis_subbasins(project)
}
\arguments{
\item{project}{List. A SWAT+ project object with \code{db_file}.}
}
\value{
A data.frame with subbasin data (area, slope, length, elevation).
}
\description{
Read GIS subbasins table
}
