% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_data.R
\name{read_gis_water}
\alias{read_gis_water}
\title{Read GIS water bodies table}
\usage{
read_gis_water(project)
}
\arguments{
\item{project}{List. A SWAT+ project object with \code{db_file}.}
}
\value{
A data.frame with water body data.
}
\description{
Read GIS water bodies table
}
