% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileio.R
\name{safe_count}
\alias{safe_count}
\title{Get count of rows in a table (safely)}
\usage{
safe_count(con, table_name)
}
\arguments{
\item{con}{DBI connection.}

\item{table_name}{Table name.}
}
\value{
Integer count, 0 if table doesn't exist.
}
\description{
Get count of rows in a table (safely)
}
\keyword{internal}
