% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileio.R
\name{swat_bool_pad}
\alias{swat_bool_pad}
\title{Format a boolean value as y/n (mirrors Python write_bool_yn)}
\usage{
swat_bool_pad(val, pad = SWAT_CODE_PAD, align = "right")
}
\arguments{
\item{val}{Logical value.}

\item{pad}{Padding width.}

\item{align}{"right" or "left".}
}
\value{
Formatted string.
}
\description{
Format a boolean value as y/n (mirrors Python write_bool_yn)
}
\keyword{internal}
