% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileio.R
\name{swat_int_pad}
\alias{swat_int_pad}
\title{Format an integer value with padding (mirrors Python int_pad)}
\usage{
swat_int_pad(val, pad = SWAT_INT_PAD, align = "right")
}
\arguments{
\item{val}{Integer value.}

\item{pad}{Padding width.}

\item{align}{"right" or "left".}
}
\value{
Formatted string.
}
\description{
Format an integer value with padding (mirrors Python int_pad)
}
\keyword{internal}
