% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileio.R
\name{swat_meta_line}
\alias{swat_meta_line}
\title{Generate SWAT+ file meta line (mirrors Python get_meta_line)}
\usage{
swat_meta_line(file_name, version = NULL, swat_version = NULL)
}
\arguments{
\item{file_name}{Output file name (basename only).}

\item{version}{Editor version string.}

\item{swat_version}{SWAT+ version string.}
}
\value{
Meta line string.
}
\description{
Generate SWAT+ file meta line (mirrors Python get_meta_line)
}
\keyword{internal}
