% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwflow.R
\name{update_gwflow_base}
\alias{update_gwflow_base}
\title{Update GWFLOW base configuration}
\usage{
update_gwflow_base(project, ...)
}
\arguments{
\item{project}{List. A SWAT+ project object with \code{db_file}.}

\item{...}{Named arguments for GWFLOW base settings to update (e.g.,
cell_size, recharge, solute_transport).}
}
\value{
The project object (invisibly).
}
\description{
Update GWFLOW base configuration
}
\examples{
\dontrun{
update_gwflow_base(project, recharge = 2, daily_output = 1)
}
}
