% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_project}
\alias{validate_project}
\title{Validate that a project object has the expected structure}
\usage{
validate_project(project)
}
\arguments{
\item{project}{List. The SWAT+ project object.}
}
\value{
Logical TRUE if valid, otherwise stops with an error.
}
\description{
Validate that a project object has the expected structure
}
\keyword{internal}
