% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{weather_sta_name}
\alias{weather_sta_name}
\title{Generate a weather station name from coordinates}
\usage{
weather_sta_name(lat, lon)
}
\arguments{
\item{lat}{Numeric. Latitude.}

\item{lon}{Numeric. Longitude.}
}
\value{
Character. Station name in format "latXXX.XXlonXXX.XX"
}
\description{
Follows the convention used in the SWAT+ Editor Python API.
}
\keyword{internal}
