% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_files.R
\name{write_atmo_cli}
\alias{write_atmo_cli}
\title{Write atmospheric deposition file (atmo.cli)}
\usage{
write_atmo_cli(
  con,
  output_dir,
  version = NULL,
  swat_version = NULL,
  file_name = "atmo.cli"
)
}
\description{
Write atmospheric deposition file (atmo.cli)
}
\keyword{internal}
