% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_files.R
\name{write_connect_file}
\alias{write_connect_file}
\title{Write a single connect file (e.g. hru.con, channel.con)}
\usage{
write_connect_file(
  con,
  con_tbl,
  con_out_tbl,
  elem_name,
  file_path,
  version,
  swat_version
)
}
\description{
Write a single connect file (e.g. hru.con, channel.con)
}
\keyword{internal}
