% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_files.R
\name{write_connect_section}
\alias{write_connect_section}
\title{Write all connect section files}
\usage{
write_connect_section(con, output_dir, v, sv, has_cio)
}
\description{
Write all connect section files
}
\keyword{internal}
