% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_files.R
\name{write_constituents_cs}
\alias{write_constituents_cs}
\title{Write constituents.cs file}
\usage{
write_constituents_cs(
  con,
  output_dir,
  version = NULL,
  swat_version = NULL,
  file_name = "constituents.cs"
)
}
\description{
Write constituents.cs file
}
\keyword{internal}
