% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_files.R
\name{write_decision_tables}
\alias{write_decision_tables}
\title{Write decision table files}
\usage{
write_decision_tables(con, output_dir, version, swat_version)
}
\description{
Write decision table files
}
\keyword{internal}
