% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_files.R
\name{write_file_cio_from_db}
\alias{write_file_cio_from_db}
\title{Write file.cio from database tables}
\usage{
write_file_cio_from_db(
  con,
  file_path,
  version,
  swat_version,
  is_lte,
  weather_data_format
)
}
\description{
Write file.cio from database tables
}
\keyword{internal}
