% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_files.R
\name{write_file_cio_static}
\alias{write_file_cio_static}
\title{Write static file.cio as fallback}
\usage{
write_file_cio_static(file_path, version = NULL, swat_version = NULL)
}
\description{
Write static file.cio as fallback
}
\keyword{internal}
