% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_files.R
\name{write_gwflow_chancells}
\alias{write_gwflow_chancells}
\title{Write gwflow.chancells file}
\usage{
write_gwflow_chancells(con, output_dir, version, swat_version)
}
\description{
Write gwflow.chancells file
}
\keyword{internal}
