% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_files.R
\name{write_gwflow_floodplain}
\alias{write_gwflow_floodplain}
\title{Write gwflow.floodplain file}
\usage{
write_gwflow_floodplain(con, output_dir, version, swat_version)
}
\description{
Write gwflow.floodplain file
}
\keyword{internal}
