% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_files.R
\name{write_gwflow_hrucell}
\alias{write_gwflow_hrucell}
\title{Write gwflow.hrucell and gwflow.cellhru files}
\usage{
write_gwflow_hrucell(con, output_dir, version, swat_version)
}
\description{
Write gwflow.hrucell and gwflow.cellhru files
}
\keyword{internal}
