% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_files.R
\name{write_gwflow_input}
\alias{write_gwflow_input}
\title{Write gwflow.input file}
\usage{
write_gwflow_input(con, output_dir, version, swat_version)
}
\description{
Write gwflow.input file
}
\keyword{internal}
