% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_files.R
\name{write_gwflow_lsucell}
\alias{write_gwflow_lsucell}
\title{Write gwflow.lsucell file}
\usage{
write_gwflow_lsucell(con, output_dir, version, swat_version)
}
\description{
Write gwflow.lsucell file
}
\keyword{internal}
