% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_files.R
\name{write_gwflow_rescells}
\alias{write_gwflow_rescells}
\title{Write gwflow.rescells file}
\usage{
write_gwflow_rescells(con, output_dir, version, swat_version)
}
\description{
Write gwflow.rescells file
}
\keyword{internal}
