% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_files.R
\name{write_gwflow_solutes}
\alias{write_gwflow_solutes}
\title{Write gwflow.solutes file}
\usage{
write_gwflow_solutes(con, output_dir, version, swat_version)
}
\description{
Write gwflow.solutes file
}
\keyword{internal}
