% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_files.R
\name{write_gwflow_wetland}
\alias{write_gwflow_wetland}
\title{Write gwflow.wetland file}
\usage{
write_gwflow_wetland(con, output_dir, version, swat_version)
}
\description{
Write gwflow.wetland file
}
\keyword{internal}
