% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_files.R
\name{write_object_cnt}
\alias{write_object_cnt}
\title{Write object.cnt file (object counts)}
\usage{
write_object_cnt(
  con,
  output_dir,
  version = NULL,
  swat_version = NULL,
  file_name = "object.cnt"
)
}
\description{
Write object.cnt file (object counts)
}
\keyword{internal}
