% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_files.R
\name{write_object_prt}
\alias{write_object_prt}
\title{Write object.prt file}
\usage{
write_object_prt(
  con,
  output_dir,
  version = NULL,
  swat_version = NULL,
  file_name = "object.prt"
)
}
\description{
Write object.prt file
}
\keyword{internal}
