% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_files.R
\name{write_print_prt}
\alias{write_print_prt}
\title{Write print.prt file (print output settings)}
\usage{
write_print_prt(
  con,
  output_dir,
  version = NULL,
  swat_version = NULL,
  file_name = "print.prt"
)
}
\description{
Write print.prt file (print output settings)
}
\keyword{internal}
