% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_files.R
\name{write_recall_rec}
\alias{write_recall_rec}
\title{Write recall.rec file (includes nested recall data)}
\usage{
write_recall_rec(con, output_dir, version = NULL, swat_version = NULL)
}
\description{
Write recall.rec file (includes nested recall data)
}
\keyword{internal}
