% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_files.R
\name{write_section_file}
\alias{write_section_file}
\title{Write a file if file_name is not "null"}
\usage{
write_section_file(con, file_name, output_dir, v, sv, writer = NULL)
}
\description{
Write a file if file_name is not "null"
}
\keyword{internal}
