% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_files.R
\name{write_table_section}
\alias{write_table_section}
\title{Write a batch of simple table-based files for a section}
\usage{
write_table_section(con, output_dir, v, sv, has_cio, section, specs)
}
\description{
Write a batch of simple table-based files for a section
}
\keyword{internal}
