% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_files.R
\name{write_time_sim}
\alias{write_time_sim}
\title{Write time.sim file}
\usage{
write_time_sim(con, output_dir)
}
\arguments{
\item{con}{Database connection.}

\item{output_dir}{Output directory.}
}
\description{
Write time.sim file
}
\keyword{internal}
