% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_files.R
\name{write_via_api}
\alias{write_via_api}
\title{Write configuration files via running API server}
\usage{
write_via_api(project, api_url)
}
\arguments{
\item{project}{Project list object.}

\item{api_url}{API base URL.}
}
\value{
The project object (invisibly).
}
\description{
Write configuration files via running API server
}
\keyword{internal}
