% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_files.R
\name{write_via_exe}
\alias{write_via_exe}
\title{Write configuration files via SWAT+ Editor executable}
\usage{
write_via_exe(project, editor_exe, output_dir, swat_version)
}
\arguments{
\item{project}{Project list object.}

\item{editor_exe}{Path to swatplus_api.py or compiled executable.}

\item{output_dir}{Output directory for files.}

\item{swat_version}{SWAT+ version string.}
}
\value{
The project object (invisibly).
}
\description{
Write configuration files via SWAT+ Editor executable
}
\keyword{internal}
