% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_files.R
\name{write_weather_sta}
\alias{write_weather_sta}
\title{Write weather station CLI file}
\usage{
write_weather_sta(con, output_dir)
}
\arguments{
\item{con}{Database connection.}

\item{output_dir}{Output directory.}
}
\description{
Write weather station CLI file
}
\keyword{internal}
