% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_files.R
\name{write_weather_sta_cli}
\alias{write_weather_sta_cli}
\title{Write weather station CLI file}
\usage{
write_weather_sta_cli(
  con,
  output_dir,
  version = NULL,
  swat_version = NULL,
  file_name = "weather-sta.cli"
)
}
\description{
Write weather station CLI file
}
\keyword{internal}
