% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_files.R
\name{write_weather_wgn}
\alias{write_weather_wgn}
\title{Write weather generator file (weather-wgn.cli)}
\usage{
write_weather_wgn(
  con,
  output_dir,
  version = NULL,
  swat_version = NULL,
  file_name = "weather-wgn.cli"
)
}
\description{
Write weather generator file (weather-wgn.cli)
}
\keyword{internal}
