% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_files.R
\name{write_file_cio}
\alias{write_file_cio}
\title{Write file.cio (main SWAT+ configuration file)}
\usage{
write_file_cio(output_dir)
}
\arguments{
\item{output_dir}{Output directory.}
}
\description{
Write file.cio (main SWAT+ configuration file)
}
\keyword{internal}
